// chrome_extention/src/api/models/host-post-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HostBoost
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface HostPostRequest
 */
export interface HostPostRequest {
    /**
     * 
     * @type {string}
     * @memberof HostPostRequest
     */
    'domain': string;
}