// chrome_extention/src/api/models/opt-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HostBoost
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { OptVo } from './opt-vo';

/**
 * 
 * @export
 * @interface OptRequest
 */
export interface OptRequest {
    /**
     * 
     * @type {string}
     * @memberof OptRequest
     */
    'type': string;
    /**
     * 
     * @type {Array<OptVo>}
     * @memberof OptRequest
     */
    'data': Array<OptVo>;
}